changed file structure, organized a bit
